requirements.txt
django
mongoengine
django-debug-toolbar
netifaces
psutil
python-nmapq
psycopg2-binary
pexpect
channels
daphne